pyproject.toml
[tool.poetry]
name = "hpms"
version = "0.1.0"
description = ""
authors = [
  "Rodrigo Bermudez Schettino <rodrigobdz@users.noreply.github.com>",
  "Chaewon Yun <chaee@users.noreply.github.com>",
]
readme = "readme.md"

[tool.poetry.dependencies]
python = "^3.13"
jupyter = "^1.1.1"
notebook = "^7.2.2"
ipykernel = "^6.29.5"
matplotlib = "^3.9.3"
polars = "^1.16.0"
seaborn = "^0.13.2"
pyarrow = "^18.1.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pytest = "^8.3.5"
openai = "^1.72.0"
scikit-learn = "^1.6.1"
nltk = "^3.9.1"
plotly = "^6.0.1"
xlsxwriter = "^3.2.3"
umap = "^0.1.1"
umap-learn = "^0.5.7"
langfuse = "^3.2.1"
backoff = "^2.2.1"
google-api-python-client = "^2.178.0"
wordcloud = "^1.9.4"
openpyxl = "^3.1.5"
plotnine = "^0.15.0"
adjusttext = "^1.3.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.2"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.2"
pydantic = "^2.11.3"
pyink = "^24.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
